refactor: keep profile edit groups in profile manager
